+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ v0.61.1 (2026-05-25)
 --------------------
 Contributors to this version: Pascal Bourgault (:user:`aulemahal`).
 
+New indicators and features
+^^^^^^^^^^^^^^^^^^^^^^^^^^^
+* Run length indices now support a quantile reducer.
+
 Bug fixes
 ^^^^^^^^^
 * Fix conversion error with ``xc.units.rate2amount`` and ``xc.units.amount2rate`` when ``sampling_rate_from_coord=True`` or sampling frequency is monthly or coarser and time coordinate is `cftime`-based. Previous results were 1000x too small. (:pull:`2357`).